Break apart common_types mod
