+++ crates/common/src/api/mod.rs
@@ -1,0 +1,4 @@
+pub mod hallucination;
+pub mod open_ai;
+pub mod prompt_guard;
+pub mod zero_shot;
