+++ crates/prompt_gateway/src/embeddings.rs
@@ -1,0 +1,5 @@
+#[derive(Debug, Clone, Copy, Hash, PartialEq, Eq)]
+pub enum EmbeddingType {
+    Name,
+    Description,
+}
